add endpoint retornarBicicleta e seus testes

+++ src/main/java/com/bikeunirio/bicicletario/equipamento/repository/BicicletaRepository.java
@@ -3,6 +3,6 @@ package com.bikeunirio.bicicletario.equipamento.repository;
 import com.bikeunirio.bicicletario.equipamento.entity.Bicicleta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BicicletaRepository extends JpaRepository<Bicicleta, Integer> {
+public interface BicicletaRepository extends JpaRepository<Bicicleta, Long> {
 
 }
